+++ src/main/java/ro/webdata/normalization/timespan/ro/TimespanUtils.java
@@ -41,49 +41,49 @@ public class TimespanUtils {
      * </ol>
      * @param original The original value taken from "lido:displayDate" record
      */
-    public static TimespanModel prepareTimespanModel(String original) {
+    public static TimespanModel prepareTimespanModel(String original, boolean historicalOnly) {
         String value = StringUtils.stripAccents(original);
         TimespanModel timespanModel = new TimespanModel(value);
-        updateMatchedValues(original, timespanModel, UnknownRegex.UNKNOWN, TimespanType.UNKNOWN);
+        updateMatchedValues(original, timespanModel, historicalOnly, UnknownRegex.UNKNOWN, TimespanType.UNKNOWN);
 
-        updateMatchedValues(original, timespanModel, DateRegex.DATE_DMY_INTERVAL, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, DateRegex.DATE_YMD_INTERVAL, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, ShortDateRegex.DATE_MY_INTERVAL, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, DateRegex.DATE_DMY_OPTIONS, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, DateRegex.DATE_YMD_OPTIONS, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, ShortDateRegex.DATE_MY_OPTIONS, TimespanType.DATE);
-        updateMatchedValues(original, timespanModel, LongDateRegex.LONG_DATE_OPTIONS, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, DateRegex.DATE_DMY_INTERVAL, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, DateRegex.DATE_YMD_INTERVAL, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, ShortDateRegex.DATE_MY_INTERVAL, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, DateRegex.DATE_DMY_OPTIONS, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, DateRegex.DATE_YMD_OPTIONS, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, ShortDateRegex.DATE_MY_OPTIONS, TimespanType.DATE);
+        updateMatchedValues(original, timespanModel, historicalOnly, LongDateRegex.LONG_DATE_OPTIONS, TimespanType.DATE);
 
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.CENTURY_INTERVAL, TimespanType.CENTURY);
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.CENTURY_OPTIONS, TimespanType.CENTURY);
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.MILLENNIUM_INTERVAL, TimespanType.MILLENNIUM);
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.MILLENNIUM_OPTIONS, TimespanType.MILLENNIUM);
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.OTHER_CENTURY_ROMAN_INTERVAL, TimespanType.CENTURY);
-        updateMatchedValues(original, timespanModel, TimePeriodRegex.OTHER_CENTURY_ROMAN_OPTIONS, TimespanType.CENTURY);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.CENTURY_INTERVAL, TimespanType.CENTURY);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.CENTURY_OPTIONS, TimespanType.CENTURY);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.MILLENNIUM_INTERVAL, TimespanType.MILLENNIUM);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.MILLENNIUM_OPTIONS, TimespanType.MILLENNIUM);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.OTHER_CENTURY_ROMAN_INTERVAL, TimespanType.CENTURY);
+        updateMatchedValues(original, timespanModel, historicalOnly, TimePeriodRegex.OTHER_CENTURY_ROMAN_OPTIONS, TimespanType.CENTURY);
         for (int i = 0; i < AgeRegex.AGE_OPTIONS.length; i++) {
-            updateMatchedValues(original, timespanModel, AgeRegex.AGE_OPTIONS[i], TimespanType.EPOCH);
+            updateMatchedValues(original, timespanModel, historicalOnly, AgeRegex.AGE_OPTIONS[i], TimespanType.EPOCH);
         }
 
-        updateMatchedValues(original, timespanModel, DatelessRegex.DATELESS, TimespanType.UNKNOWN);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.AFTER_INTERVAL, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.BEFORE_INTERVAL, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.APPROX_AGES_INTERVAL, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.AFTER, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.BEFORE, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, DatelessRegex.DATELESS, TimespanType.UNKNOWN);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.AFTER_INTERVAL, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.BEFORE_INTERVAL, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.APPROX_AGES_INTERVAL, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.AFTER, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.BEFORE, TimespanType.YEAR);
 
-        updateMatchedValues(original, timespanModel, YearRegex.YEAR_INTERVAL, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, InaccurateYearRegex.APPROX_AGES_OPTIONS, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, YearRegex.YEAR_3_4_DIGITS_SPECIAL_INTERVAL, TimespanType.YEAR);
-        updateMatchedValues(original, timespanModel, YearRegex.YEAR_OPTIONS, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, YearRegex.YEAR_INTERVAL, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, InaccurateYearRegex.APPROX_AGES_OPTIONS, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, YearRegex.YEAR_3_4_DIGITS_SPECIAL_INTERVAL, TimespanType.YEAR);
+        updateMatchedValues(original, timespanModel, historicalOnly, YearRegex.YEAR_OPTIONS, TimespanType.YEAR);
         // This call need to be made after all the years processing !!!
-        updateMatchedValues(original, timespanModel, YearRegex.UNKNOWN_YEARS, TimespanType.UNKNOWN);
+        updateMatchedValues(original, timespanModel, historicalOnly, YearRegex.UNKNOWN_YEARS, TimespanType.UNKNOWN);
 
         return timespanModel;
     }
 
     //TODO: "1/2 mil. 5 - sec. I al mil. 4 a.Chr."
     //TODO: "2 a.chr - 14 p.chr"
-    private static void updateMatchedValues(String original, TimespanModel timespanModel, String regex, String matchedType) {
+    private static void updateMatchedValues(String original, TimespanModel timespanModel, boolean historicalOnly, String regex, String matchedType) {
         String initialValue = timespanModel.getResidualValue();
         initialValue = TimeSanitizeUtils.sanitizeValue(initialValue, regex);
         String residualValue = initialValue
@@ -99,7 +99,7 @@ public class TimespanUtils {
             }
 
             matchedValue = matchedValue.trim();
-            TimePeriodModel timePeriod = prepareTimePeriodModel(original, TimeUtils.normalizeChristumNotation(matchedValue), regex);
+            TimePeriodModel timePeriod = prepareTimePeriodModel(original, TimeUtils.normalizeChristumNotation(matchedValue), historicalOnly, regex);
             String matchedItems = timePeriod.toString();
 
             Map<String, DBpediaModel> edges = TimespanUtils.prepareEdges(timePeriod, matchedType, matchedValue);
@@ -139,18 +139,20 @@ public class TimespanUtils {
      * Ensure the right format for date-like, year-like, centuries and millenniums values
      * @param original The original value
      * @param value The value subjected to the replacement process (part of the original value)
+     * @param historicalOnly Flag which specifies whether the Framework will only handle historical
+     *                       dates (future dates will be ignored)
      * @param regex A regular expression
      * @return The formatted value
      */
-    private static TimePeriodModel prepareTimePeriodModel(String original, String value, String regex) {
-        TimePeriodModel prepared = prepareAges(original, value, regex);
+    private static TimePeriodModel prepareTimePeriodModel(String original, String value, boolean historicalOnly, String regex) {
+        TimePeriodModel prepared = prepareAges(original, value, historicalOnly, regex);
 
         if (prepared == null) {
-            prepared = prepareDateTime(original, value, regex);
+            prepared = prepareDateTime(original, value, historicalOnly, regex);
         }
 
         if (prepared == null) {
-            prepared = preparePeriod(original, value, regex);
+            prepared = preparePeriod(original, value, historicalOnly, regex);
         }
 
         return prepared;
@@ -160,10 +162,12 @@ public class TimespanUtils {
      * Ensure the right format for year-like value
      * @param original The original value
      * @param value The value subjected to the replacement process (part of the original value)
+     * @param historicalOnly Flag which specifies whether the Framework will only handle historical
+     *                       dates (future dates will be ignored)
      * @param regex A regular expression
      * @return The formatted value
      */
-    private static TimePeriodModel prepareAges(String original, String value, String regex) {
+    private static TimePeriodModel prepareAges(String original, String value, boolean historicalOnly, String regex) {
         switch (regex) {
             case InaccurateYearRegex.AFTER:
             case InaccurateYearRegex.AFTER_INTERVAL:
@@ -171,15 +175,15 @@ public class TimespanUtils {
             case InaccurateYearRegex.APPROX_AGES_OPTIONS:
             case InaccurateYearRegex.BEFORE:
             case InaccurateYearRegex.BEFORE_INTERVAL:
-                return new InaccurateYearModel(original, value);
+                return new InaccurateYearModel(original, value, historicalOnly);
             case DatelessRegex.DATELESS:
             case YearRegex.UNKNOWN_YEARS:
             case UnknownRegex.UNKNOWN:
-                return new DatelessModel(original, value);
+                return new DatelessModel(original, value, historicalOnly);
             case YearRegex.YEAR_INTERVAL:
             case YearRegex.YEAR_3_4_DIGITS_SPECIAL_INTERVAL:
             case YearRegex.YEAR_OPTIONS:
-                return new YearModel(original, value);
+                return new YearModel(original, value, historicalOnly);
             default:
                 return null;
         }
@@ -189,22 +193,24 @@ public class TimespanUtils {
      * Ensure the right format for date-like value
      * @param original The original value
      * @param value The value subjected to the replacement process (part of the original value)
+     * @param historicalOnly Flag which specifies whether the Framework will only handle historical
+     *                       dates (future dates will be ignored)
      * @param regex A regular expression
      * @return The formatted value
      */
-    private static TimePeriodModel prepareDateTime(String original, String value, String regex) {
+    private static TimePeriodModel prepareDateTime(String original, String value, boolean historicalOnly, String regex) {
         switch (regex) {
             case DateRegex.DATE_DMY_INTERVAL:
             case DateRegex.DATE_DMY_OPTIONS:
-                return new DateModel(original, value, TimeUtils.DMY_PLACEHOLDER);
+                return new DateModel(original, value, TimeUtils.DMY_PLACEHOLDER, historicalOnly);
             case DateRegex.DATE_YMD_INTERVAL:
             case DateRegex.DATE_YMD_OPTIONS:
-                return new DateModel(original, value, TimeUtils.YMD_PLACEHOLDER);
+                return new DateModel(original, value, TimeUtils.YMD_PLACEHOLDER, historicalOnly);
             case ShortDateRegex.DATE_MY_INTERVAL:
             case ShortDateRegex.DATE_MY_OPTIONS:
-                return new ShortDateModel(original, value, TimeUtils.MY_PLACEHOLDER);
+                return new ShortDateModel(original, value, TimeUtils.MY_PLACEHOLDER, historicalOnly);
             case LongDateRegex.LONG_DATE_OPTIONS:
-                return new LongDateModel(original, value);
+                return new LongDateModel(original, value, historicalOnly);
             default:
                 return null;
         }
@@ -214,19 +220,21 @@ public class TimespanUtils {
      * Ensure the right format for centuries and millenniums
      * @param original The original value
      * @param value The value subjected to the replacement process (part of the original value)
+     * @param historicalOnly Flag which specifies whether the Framework will only handle historical
+     *                       dates (future dates will be ignored)
      * @param regex A regular expression
      * @return The list of formatted values
      */
-    private static TimePeriodModel preparePeriod(String original, String value, String regex) {
+    private static TimePeriodModel preparePeriod(String original, String value, boolean historicalOnly, String regex) {
         switch (regex) {
             case TimePeriodRegex.CENTURY_INTERVAL:
             case TimePeriodRegex.CENTURY_OPTIONS:
             case TimePeriodRegex.OTHER_CENTURY_ROMAN_INTERVAL:
             case TimePeriodRegex.OTHER_CENTURY_ROMAN_OPTIONS:
-                return new CenturyModel(original, value);
+                return new CenturyModel(original, value, historicalOnly);
             case TimePeriodRegex.MILLENNIUM_INTERVAL:
             case TimePeriodRegex.MILLENNIUM_OPTIONS:
-                return new MillenniumModel(original, value);
+                return new MillenniumModel(original, value, historicalOnly);
             case AgeRegex.AURIGNACIAN_CULTURE:
             case AgeRegex.BRONZE_AGE:
             case AgeRegex.CHALCOLITHIC_AGE:
@@ -244,7 +252,7 @@ public class TimespanUtils {
             case AgeRegex.ROMAN_EMPIRE_AGE:
             case AgeRegex.WW_I_PERIOD:
             case AgeRegex.WW_II_PERIOD:
-                return new AgeModel(original, value, regex);
+                return new AgeModel(original, value, historicalOnly, regex);
             default:
                 return null;
         }