+++ src/main/java/ro/webdata/normalization/timespan/ro/model/imprecise/DatelessModel.java
@@ -5,7 +5,7 @@ import ro.webdata.normalization.timespan.ro.model.TimePeriodModel;
 public class DatelessModel extends TimePeriodModel {
     private String value;
 
-    public DatelessModel(String original, String value) {
+    public DatelessModel(String original, String value, boolean historicalOnly) {
         this.value = value;
     }
 }