+++ src/main/java/ro/webdata/normalization/timespan/ro/model/imprecise/InaccurateYearModel.java
@@ -12,11 +12,11 @@ import ro.webdata.normalization.timespan.ro.regex.TimespanRegex;
 public class InaccurateYearModel extends TimePeriodModel {
     private static final String REGEX_NON_DIGIT = "[^\\d]";
 
-    public InaccurateYearModel(String original, String value) {
-        setDateModel(original, value);
+    public InaccurateYearModel(String original, String value, boolean historicalOnly) {
+        setDateModel(original, value, historicalOnly);
     }
 
-    private void setDateModel(String original, String value) {
+    private void setDateModel(String original, String value, boolean historicalOnly) {
         String[] intervalValues = value.split(TimespanRegex.REGEX_INTERVAL_DELIMITER);
 
         if (intervalValues.length == 2) {
@@ -25,24 +25,24 @@ public class InaccurateYearModel extends TimePeriodModel {
             String startValue = TimeUtils.clearChristumNotation(intervalValues[0]);
             String endValue = TimeUtils.clearChristumNotation(intervalValues[1]);
 
-            setDate(original, endValue, TimeUtils.END_PLACEHOLDER);
-            setDate(original, startValue, TimeUtils.START_PLACEHOLDER);
+            setDate(original, endValue, TimeUtils.END_PLACEHOLDER, historicalOnly);
+            setDate(original, startValue, TimeUtils.START_PLACEHOLDER, historicalOnly);
         } else {
             setEra(original, value, value, true);
 
             String preparedValue = TimeUtils.clearChristumNotation(value);
 
-            setDate(original, preparedValue, TimeUtils.END_PLACEHOLDER);
-            setDate(original, preparedValue, TimeUtils.START_PLACEHOLDER);
+            setDate(original, preparedValue, TimeUtils.END_PLACEHOLDER, historicalOnly);
+            setDate(original, preparedValue, TimeUtils.START_PLACEHOLDER, historicalOnly);
         }
     }
 
-    private void setDate(String original, String value, String position) {
+    private void setDate(String original, String value, String position, boolean historicalOnly) {
         String year = value
                 .replaceAll(REGEX_NON_DIGIT, Const.EMPTY_VALUE_PLACEHOLDER);
 
-        setMillennium(original, year, position);
-        setCentury(original, year, position);
-        setYear(original, year, position);
+        setMillennium(original, year, position, historicalOnly);
+        setCentury(original, year, position, historicalOnly);
+        setYear(original, year, position, historicalOnly);
     }
 }